0.31.3.0: Update code style, line length, member order, etc. (#111)
* Update Microsoft.NET.Test.Sdk

* Code style fixes. Mostly use of null check throw pattern. But also ordering of naming members in unit tests, and a few other details

* Remove interval support in impurity metrics. These were developed for decision tree learning originally, but was replaced by alternative implementation. Also remove InCustomDictionary, which was introduced for performance reasons. This is no longer a priority since the Impurity metrics are no longer used in the learning algorithms

* Remove DiscreteTargetMeanErrorRegressionMetric and RocAucRegressionMetric, since these are not "classic" metric. Fix line lengths

+++ src/SharpLearning.GradientBoost/Loss/GradientBoostSquaredLoss.cs
@@ -1,9 +1,4 @@
 ﻿using SharpLearning.GradientBoost.GBMDecisionTree;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SharpLearning.GradientBoost.Loss
 {
@@ -70,8 +65,8 @@ namespace SharpLearning.GradientBoost.Loss
                 }
             }
 
-            splitInfo.Cost = splitInfo.SumOfSquares - (splitInfo.Sum * splitInfo.Sum / (double)splitInfo.Samples);
-            splitInfo.BestConstant = splitInfo.Sum / (double)splitInfo.Samples;
+            splitInfo.Cost = splitInfo.SumOfSquares - (splitInfo.Sum * splitInfo.Sum / splitInfo.Samples);
+            splitInfo.BestConstant = splitInfo.Sum / splitInfo.Samples;
 
             return splitInfo;
         }
@@ -94,7 +89,8 @@ namespace SharpLearning.GradientBoost.Loss
         /// <param name="predictions"></param>
         /// <param name="residuals"></param>
         /// <param name="inSample"></param>
-        public void UpdateResiduals(double[] targets, double[] predictions, double[] residuals, bool[] inSample)
+        public void UpdateResiduals(double[] targets, double[] predictions, 
+            double[] residuals, bool[] inSample)
         {
             for (int i = 0; i < residuals.Length; i++)
             {
@@ -112,15 +108,16 @@ namespace SharpLearning.GradientBoost.Loss
         /// <param name="right"></param>
         /// <param name="target"></param>
         /// <param name="residual"></param>
-        public void UpdateSplitConstants(ref GBMSplitInfo left, ref GBMSplitInfo right, double target, double residual)
+        public void UpdateSplitConstants(ref GBMSplitInfo left, ref GBMSplitInfo right, 
+            double target, double residual)
         {
             var residual2 = residual * residual;
 
             left.Samples++;
             left.Sum += residual;
             left.SumOfSquares += residual2;
-            left.Cost = left.SumOfSquares - (left.Sum * left.Sum / (double)left.Samples);
-            left.BestConstant = left.Sum / (double)left.Samples;
+            left.Cost = left.SumOfSquares - (left.Sum * left.Sum / left.Samples);
+            left.BestConstant = left.Sum / left.Samples;
             
             // Alternative update but gives slightly different results
             //var leftSamplesInv =  1.0 / left.Samples;
@@ -134,8 +131,8 @@ namespace SharpLearning.GradientBoost.Loss
             right.Samples--;
             right.Sum -= residual;
             right.SumOfSquares -= residual2;
-            right.Cost = right.SumOfSquares - (right.Sum * right.Sum / (double)right.Samples);
-            right.BestConstant = right.Sum / (double)right.Samples;
+            right.Cost = right.SumOfSquares - (right.Sum * right.Sum / right.Samples);
+            right.BestConstant = right.Sum / right.Samples;
 
             // Alternative update but gives slightly different results
             //var rightSamplesInv = 1.0 / right.Samples;
@@ -154,7 +151,8 @@ namespace SharpLearning.GradientBoost.Loss
         /// <param name="predictions"></param>
         /// <param name="inSample"></param>
         /// <returns></returns>
-        public double UpdatedLeafValue(double currentLeafValue, double[] targets, double[] predictions, bool[] inSample)
+        public double UpdatedLeafValue(double currentLeafValue, double[] targets, 
+            double[] predictions, bool[] inSample)
         {
             // no updates needed for square loss
             return currentLeafValue;
